Added new features
Custom exception handler, Room controller,service and DAO have been added. Was created many-to-many table "Booking".

+++ conference-room-booking/src/main/java/com/conferencebooking/conferenceroombooking/restcontroller/UserRestController.java
@@ -1,57 +1,76 @@
 package com.conferencebooking.conferenceroombooking.restcontroller;
 
 import com.conferencebooking.conferenceroombooking.entity.User;
-
-import com.conferencebooking.conferenceroombooking.service.UserService;
-import org.hibernate.MappingException;
-import org.hibernate.TransactionException;
+import com.conferencebooking.conferenceroombooking.exceptions.AccessDeniedException;
+import com.conferencebooking.conferenceroombooking.model.UserDTO;
+import com.conferencebooking.conferenceroombooking.service.user.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.PropertySource;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.server.ResponseStatusException;
+
 
 import java.util.List;
 
 @RestController
 @RequestMapping("/api")
+@PropertySource("classpath:admin-file.properties")
 public class UserRestController {
 
     private UserService userService;
 
+    @Value("${adminkey}")
+    private String securityKey;
+
     @Autowired
-    public UserRestController(UserService theUserService ){
-        userService=theUserService;
+    public UserRestController(UserService theUserService) {
+        userService = theUserService;
     }
 
     @GetMapping("/users")
-    public List<User> getAvailableUsers(){
+    public List<UserDTO> getAvailableUsers() {
         return userService.getAvailableUsers();
     }
 
-    @PutMapping("/user/{login}")
-    public String updateUser(@PathVariable("login") String login,
-                             @RequestParam(name = "name",required = false)String name,
-                             @RequestParam(name = "surname",required = false)String surname,
-                             @RequestParam(name = "password",required = false)String pass ){
 
-
-        userService.updateTheUser(login, name, surname, pass);
-        return "User has been updated!";
+    @PutMapping("/user/{adminkey}")
+    public ResponseEntity<?>updateUser(@PathVariable("adminkey") String adminKey, @RequestBody User user){
+        if (adminKey.equals(securityKey)) {
+            userService.updateTheUser(user);
+            return new ResponseEntity<>("Update has been completed!", HttpStatus.OK);
+        } else {
+            throw new AccessDeniedException();
+        }
     }
 
-    @PostMapping("/user")
-    public User addUser(@RequestBody User theUser){
+    @PostMapping("/user/{adminkey}")
+    public ResponseEntity<?> addUser(@PathVariable("adminkey") String adminKey, @Validated @RequestBody User theUser) {
 
-        theUser.setId(0);
-        userService.saveTheUser(theUser);
+        if (adminKey.equals(securityKey)) {
 
-        return theUser;
+            userService.saveTheUser(theUser);
+            return new ResponseEntity<>(theUser, HttpStatus.OK);
+
+        } else {
+            throw new AccessDeniedException();
+        }
 
     }
 
-    @DeleteMapping("/user/{login}")
-    public void deleteUser(@PathVariable String login){
-        userService.deleteTheUser(login);
+    @DeleteMapping("/user/{adminkey}/{login}")
+    public ResponseEntity<?> deleteUser(@PathVariable("adminkey") String adminKey, @PathVariable("login") String login) {
+
+        if (adminKey.equals(securityKey)) {
+
+            userService.deleteTheUser(login);
+            return new ResponseEntity<>("User has been deleted!", HttpStatus.OK);
+
+        } else {
+            throw new AccessDeniedException();
+        }
     }
